svdbased service almost finished + similarUsers

+++ sbt/src/main/scala/odi/recommendation/service/ItemBasedService.scala
@@ -26,7 +26,6 @@ object ItemBasedService extends HttpServer {
 
   //for each item that is connectet with a user find all items from the user and calculate the similarity with the original item
   //should work with a futurepool
-  //I need purchased together otherwise all similarities are doubled because item1 -> item2 and item2 -> item1
   def getCalculateSimilarItems(path: Array[String]): Future[HttpResponse] = {
     val items: List[Item] = Items.all
     val purchasedTogether = collection.mutable.Set[(Item, Item)]() //all items that where purchased together by one or more users
@@ -65,18 +64,18 @@ object ItemBasedService extends HttpServer {
   def getCalculateUserPredictions(userId: Int, path: Array[String]): Future[HttpResponse] = {
 
     //save the items that are unknown together with the item that are known and their similarity values
-    val similarItems = collection.mutable.HashMap[Item, collection.mutable.LinkedList[(Item, Float)]]()
+    val similarItems = collection.mutable.HashMap[Item, collection.mutable.LinkedList[(Item, Double)]]()
     val allItemsUser = Items.allItemsUser(userId)
 
     for(userItem: Item <- allItemsUser;
-        similarItem: (Item, Float) <- userItem.similarItems) 
+        similarItem: (Item, Double) <- userItem.similarItems) 
     {
       if(!allItemsUser.contains(similarItem._1)) //item is unknown to the user
           if(similarItems.contains(similarItem._1)) {
-            similarItems(similarItem._1).append(collection.mutable.LinkedList[(Item, Float)]((userItem, similarItem._2)))
+            similarItems(similarItem._1).append(collection.mutable.LinkedList[(Item, Double)]((userItem, similarItem._2)))
           }
           else {
-            similarItems += (similarItem._1 -> collection.mutable.LinkedList[(Item, Float)]((userItem, similarItem._2))) 
+            similarItems += (similarItem._1 -> collection.mutable.LinkedList[(Item, Double)]((userItem, similarItem._2))) 
           }
     }
 
@@ -86,11 +85,11 @@ object ItemBasedService extends HttpServer {
   }
 
   //calculate the prediction for one item from one User by the items that he already rated
-  def calculatePrediction(userId: Int, similarItems: List[(Item, Float)]): Int = {
+  def calculatePrediction(userId: Int, similarItems: List[(Item, Double)]): Double = {
     val numerator = similarItems.map({case (item, similarity) => {
         Ratings.getByItemUser(item.id.get, userId).get.rating * similarity
       }}).sum 
-    (numerator / similarItems.map(_._2).sum).toInt
+    (numerator / similarItems.map(_._2).sum)
   }
 
 }
